perf: Create separate bundles for all icon sets
